Refactor CLI interface code to a separate module
