+++ src/lib.rs
@@ -5,3 +5,6 @@ pub mod blocking {
 pub mod config {
     pub mod block;
 }
+pub mod cli {
+    pub mod cli_runner;
+}
